src/api/requirements.txt
Flask 
mysql-connector-python
pyjwt
flask-cors
redis